+++ test/test_backends.py
@@ -14,6 +14,9 @@ from __future__ import unicode_literals
 import pytest
 
 import testinfra.backend
+from testinfra.backend.base import BaseBackend
+from testinfra.backend.base import HostSpec
+from testinfra.backend.winrm import _quote
 BACKENDS = ("ssh", "safe-ssh", "docker", "paramiko", "ansible")
 HOSTS = [backend + "://debian_jessie" for backend in BACKENDS]
 USER_HOSTS = [backend + "://user@debian_jessie" for backend in BACKENDS]
@@ -97,3 +100,28 @@ def test_docker_encoding(host):
     string = "ťēꞩƫìṇḟřặ ṧꝕèȃǩ ửƫᵮ8"
     assert host.check_output("echo %s | tee /tmp/s.txt", string) == string
     assert host.file("/tmp/s.txt").content_string.strip() == string
+
+
+@pytest.mark.parametrize('hostspec,expected', [
+    ('u:P@h:p', HostSpec('h', 'p', 'u', 'P')),
+    ('u@h:p', HostSpec('h', 'p', 'u', None)),
+    ('u:P@h', HostSpec('h', None, 'u', 'P')),
+    ('u@h', HostSpec('h', None, 'u', None)),
+    ('h', HostSpec('h', None, None, None)),
+])
+def test_parse_hostspec(hostspec, expected):
+    assert BaseBackend.parse_hostspec(hostspec) == expected
+
+
+@pytest.mark.parametrize('arg_string,expected', [
+    (
+        'C:\\Users\\vagrant\\This Dir\\salt',
+        '"C:\\Users\\vagrant\\This Dir\\salt"'
+    ),
+    (
+        'C:\\Users\\vagrant\\AppData\\Local\\Temp\\kitchen\\etc\\salt',
+        '"C:\\Users\\vagrant\\AppData\\Local\\Temp\\kitchen\\etc\\salt"'
+    ),
+])
+def test_winrm_quote(arg_string, expected):
+    assert _quote(arg_string) == expected