+++ testinfra/modules/service.py
@@ -24,7 +24,7 @@ class Service(Module):
     - Linux: detect Systemd or Upstart, fallback to SysV
     - FreeBSD: service(1)
     - OpenBSD: ``/etc/rc.d/$name check`` for ``is_running``
-      (``is_enabled`` is not yet implemented)
+      ``rcctl ls on`` for ``is_enabled`` (only OpenBSD >= 5.8)
     - NetBSD: ``/etc/rc.d/$name onestatus`` for ``is_running``
       (``is_enabled`` is not yet implemented)
 
@@ -52,7 +52,9 @@ class Service(Module):
                 and "systemd" in host.file("/sbin/init").linked_to
             ):
                 return SystemdService
-            elif host.exists("initctl"):
+            elif (host.exists("initctl")
+                    and host.exists('status')
+                    and host.file('/etc/init').is_directory):
                 return UpstartService
             return SysvService
         elif host.system_info.type == "freebsd":
@@ -69,6 +71,24 @@ class Service(Module):
 
 class SysvService(Service):
 
+    def __init__(self, name):
+        super(SysvService, self).__init__(name)
+        self._command = None
+
+    @property
+    def _service_command(self):
+        if self._command is None:
+            if self._host.exists('service'):
+                self._command = 'service'
+            # service command may not exist in PATH for non privileged users
+            elif self._host.file('/sbin/service').exists:
+                self._command = '/sbin/service'
+            elif self._host.file('/usr/sbin/service').exists:
+                self._command = '/usr/sbin/service'
+            else:
+                raise RuntimeError('cannot find "service" command')
+        return self._command
+
     @property
     def is_running(self):
         # based on /lib/lsb/init-functions
@@ -77,7 +97,8 @@ class SysvService(Service):
         # 3: not running and pid file does not exists
         # 4: Unable to determine status
         return self.run_expect(
-            [0, 1, 3], "service %s status", self.name).rc == 0
+            [0, 1, 3], "%s %s status",
+            self._service_command, self.name).rc == 0
 
     @property
     def is_enabled(self):
@@ -154,7 +175,13 @@ class OpenBSDService(Service):
 
     @property
     def is_enabled(self):
-        raise NotImplementedError
+        if self.name in self.check_output('rcctl ls on').splitlines():
+            return True
+        if self.name in self.check_output('rcctl ls off').splitlines():
+            return False
+        raise RuntimeError(
+            "Unable to determine state of {0}. Does this service exist?"
+            .format(self.name))
 
 
 class NetBSDService(Service):